+++ apps/vite-project/.gitignore
