+++ apps/vite-project/package.json
@@ -10,7 +10,9 @@
   },
   "dependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "@washingtonpost/wpds-kitchen-sink": "1.1.0",
+    "@washingtonpost/wpds-ui-kit": "1.1.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.27",
