+++ apps/vite-v2-project/.gitignore
