+++ apps/vite-v2-project/package.json
@@ -9,7 +9,9 @@
   },
   "dependencies": {
     "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "react-dom": "^18.0.0",
+    "@washingtonpost/wpds-kitchen-sink": "1.1.0",
+    "@washingtonpost/wpds-ui-kit": "1.1.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.0",
