+++ apps/vite-v2-project/vite.config.js
