+++ build.washingtonpost.com/package.json
@@ -28,6 +28,7 @@
     "@washingtonpost/wpds-accordion": "1.1.0",
     "@washingtonpost/wpds-assets": "^1.17.0",
     "@washingtonpost/wpds-ui-kit": "1.1.0",
+    "@washingtonpost/wpds-kitchen-sink": "1.1.0",
     "fuse.js": "^6.6.2",
     "gray-matter": "^4.0.2",
     "lz-string": "^1.4.4",
