+++ ui/kit/src/index.ts
@@ -11,7 +11,6 @@ export * from "@washingtonpost/wpds-accordion";
 export * from "@washingtonpost/wpds-pagination-dots";
 export * from "@washingtonpost/wpds-avatar";
 export * from "@washingtonpost/wpds-divider";
-export * from "@washingtonpost/eslint-plugin-wpds";
 export * from "@washingtonpost/wpds-fieldset";
 export * from "@washingtonpost/wpds-radio-group";
 export * from "@washingtonpost/wpds-helper-text";
