+++ ui/kitchen-sink/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@washingtonpost/wpds-kitchen-sink",
+  "version": "1.1.0",
+  "description": "WPDS Kitchen Sink",
+  "author": "WPDS Support <wpds@washpost.com>",
+  "homepage": "https://github.com/washingtonpost/wpds-ui-kit#readme",
+  "license": "MIT",
+  "source": "src/index.ts",
+  "main": "dist/index.js",
+  "module": "dist/esm/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist",
+    "README.md",
+    "src"
+  ],
+  "sideEffects": false,
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/washingtonpost/wpds-ui-kit.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1",
+    "build": "tsup src/index.ts --loader .ts=tsx --minify --format esm,cjs --dts --sourcemap --legacy-output --external react",
+    "dev": "tsup src/index.ts --format esm,cjs --watch --dts --legacy-output --external react",
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
+  },
+  "bugs": {
+    "url": "https://github.com/washingtonpost/wpds-ui-kit/issues"
+  },
+  "devDependencies": {
+    "tsup": "5.11.13",
+    "typescript": "4.5.5"
+  },
+  "peerDependencies": {
+    "@washingtonpost/wpds-ui-kit": "*",
+    "@washingtonpost/wpds-assets": "latest",
+    "react": "^16.8.6 || ^17.0.2"
+  },
+  "dependencies": {
+    "@washingtonpost/wpds-ui-kit": "1.1.0",
+    "@washingtonpost/wpds-assets": "latest"
+  }
+}
